requirements_python.txt
fastapi
uvicorn
python-multipart
jinja2
google-genai
python-dotenv
